+++ CHANGELOG.md
@@ -31,13 +31,25 @@ Here is a template for new release sections
 - [line graph] stroke width 2 (#75)
 - [line graph] contrast colors (#75)
 - map bubble links to line plots (#75)
+- [line graph] x and y labels (#79)
+- [line graph] axis reactive with window size (#79)
+
 - when clicking on submit in "Request Model" it shows a result and possibility to send data to formspree (#52)
 - Creation of a contact form page with Formspree (#84)
 - Favicon (#86)
+- Update of `selectButton` dropdown via jQuery (#88)
+- Containment measures for a given country on the `model` template (#88)
+- footer (#95)
+- Event request result view (#95)
+- `containment_measures` endpoint (#108)
+
+
 ### Changed
 - Template for `model` (label "Request model") (#62, #52)
 - Model for calculating event risk (#11)
 - Template for `request-event-evaluation` (#11)
+- Version of jQuery (also add it as static file) (#88)
 
 ### Removed
 - Template for `selections` as well as view (duplicate of `request-event-calculation`) (#52)
+- Update of `selectButton` dropdown via d3 (#88)